Fix release-publish to also update the current manual.
That is, the vesion that appears as:

	http://cairographics.org/manual
and:
	http://cairographics.org/cairo-manual.tar.gz

It was silly that we previously required a manual step to
upload the documentation (which we regularly forgot to do)
and that it uploaded with a date in the name rather than a
version. So we just drop the old doc-publish Makefile target
now as it's just not useful anymore.