Revert "[png] Complete the idempotent read_png() -> write_png()"
This reverts commit 564d64a1323c5cbcde2dd9365ac790fe8aa1c5a6.

In hindsight, and with further discussion with Jeff Muizelaar, this
behaviour of using the stored contents from the mime-data is completely
the opposite of the users' expectations. When the user calls
cairo_surface_write_to_png(), usually in the course of debugging their
rendering code, they expect the precise contents of the surface to be
saved.